Removed repeated code
- Removed repeated code on leveldb.MergeSet and goleveldb.MergeSet implementations
- Removed repeated code on leveldb and goleveldb tests
- Fixed Reader implementation for leveldb - created a isolated reader (snapshot)
