+++ README.md
@@ -49,7 +49,7 @@ export CGO_LDFLAGS='-L <path/to/leveldb/lib>'
 go get -tags leveldb -v -u -t github.com/NeowayLabs/neosearch/...
 
 cd $GOPATH/src/github.com/NeowayLabs/neosearch
-go test -tags leveldb -v ./...
+go test -tags leveldb -v ./lib/neosearch/store/leveldb ./...
 ```
 
 ## Contributing
