+++ lib/neosearch/store/leveldb/reader.go
@@ -9,35 +9,37 @@ import (
 
 // LVDBReader is the readonly view of database. It implements the KVReader interface
 type LVDBReader struct {
-	store *LVDB
+	store    *LVDB
+	snapshot *levigo.Snapshot
 }
 
 // newReader returns a new reader
-func newReader(lvdb *LVDB) *LVDBReader {
+func newReader(store *LVDB) store.KVReader {
 	return &LVDBReader{
-		store: lvdb,
+		store:    store,
+		snapshot: store.db.NewSnapshot(),
 	}
 }
 
 // Get returns the value of the given key
 func (reader *LVDBReader) Get(key []byte) ([]byte, error) {
-	return reader.store.db.Get(reader.store.readOptions, key)
-}
-
-// GetCustom is the same as Get but enables override default read options
-func (reader *LVDBReader) GetCustom(opt *levigo.ReadOptions, key []byte) ([]byte, error) {
-	return reader.store.db.Get(opt, key)
+	options := defaultReadOptions()
+	options.SetSnapshot(reader.snapshot)
+	b, err := reader.store.db.Get(options, key)
+	options.Close()
+	return b, err
 }
 
 // GetIterator returns a new KVIterator
 func (reader *LVDBReader) GetIterator() store.KVIterator {
-	var ro = reader.store.readOptions
-
-	ro.SetFillCache(false)
-	it := reader.store.db.NewIterator(ro)
-	return LVDBIterator{it}
+	options := defaultReadOptions()
+	options.SetSnapshot(reader.snapshot)
+	it := reader.store.db.NewIterator(options)
+	options.Close()
+	return &LVDBIterator{it}
 }
 
 func (reader *LVDBReader) Close() error {
+	reader.store.db.ReleaseSnapshot(reader.snapshot)
 	return nil
 }
