+++ lib/neosearch/store/leveldb/writer.go
@@ -3,12 +3,9 @@
 package leveldb
 
 import (
-	"bytes"
-	"encoding/binary"
-	"fmt"
 	"sync"
 
-	"github.com/NeowayLabs/neosearch/lib/neosearch/utils"
+	"github.com/NeowayLabs/neosearch/lib/neosearch/store"
 	"github.com/jmhodges/levigo"
 )
 
@@ -36,74 +33,24 @@ func (w *LVDBWriter) Set(key []byte, value []byte) error {
 		return nil
 	}
 
-	return w.store.db.Put(w.store.writeOptions, key, value)
+	options := defaultWriteOptions()
+	err := w.store.db.Put(options, key, value)
+	options.Close()
+	return err
 }
 
-// SetCustom is the same as Set but enables override default write options
-func (w *LVDBWriter) SetCustom(opt *levigo.WriteOptions, key []byte, value []byte) error {
-	w.mutex.Lock()
-	defer w.mutex.Unlock()
-
-	return w.store.db.Put(opt, key, value)
+// Get returns the value of the given key
+func (w *LVDBWriter) Get(key []byte) ([]byte, error) {
+	options := defaultReadOptions()
+	b, e := w.store.db.Get(options, key)
+	options.Close()
+	return b, e
 }
 
 // MergeSet add value to a ordered set of integers stored in key. If value
 // is already on the key, than the set will be skipped.
 func (w *LVDBWriter) MergeSet(key []byte, value uint64) error {
-	var (
-		buf      *bytes.Buffer
-		err      error
-		v        uint64
-		i        uint64
-		inserted bool
-		reader   = w.store.Reader()
-	)
-
-	data, err := reader.Get(key)
-
-	if err != nil {
-		return err
-	}
-
-	if w.store.debug {
-		fmt.Printf("[INFO] %d ids == %d GB of ids\n", len(data)/8, len(data)/(1024*1024*1024))
-	}
-
-	buf = new(bytes.Buffer)
-	lenBytes := uint64(len(data))
-
-	// O(n)
-	for i = 0; i < lenBytes; i += 8 {
-		v = utils.BytesToUint64(data[i : i+8])
-
-		// returns if value is already stored
-		if v == value {
-			return nil
-		}
-
-		if value < v {
-			err = binary.Write(buf, binary.BigEndian, value)
-			if err != nil {
-				return err
-			}
-
-			inserted = true
-		}
-
-		err = binary.Write(buf, binary.BigEndian, v)
-		if err != nil {
-			return err
-		}
-	}
-
-	if lenBytes == 0 || !inserted {
-		err = binary.Write(buf, binary.BigEndian, value)
-		if err != nil {
-			return err
-		}
-	}
-
-	return w.Set(key, buf.Bytes())
+	return store.MergeSet(w, key, value, w.store.debug)
 }
 
 // Delete remove the given key
@@ -116,15 +63,10 @@ func (w *LVDBWriter) Delete(key []byte) error {
 		return nil
 	}
 
-	return w.store.db.Delete(w.store.writeOptions, key)
-}
-
-// DeleteCustom is the same as Delete but enables override default write options
-func (w *LVDBWriter) DeleteCustom(opt *levigo.WriteOptions, key []byte) error {
-	w.mutex.Lock()
-	defer w.mutex.Unlock()
-
-	return w.store.db.Delete(opt, key)
+	options := defaultWriteOptions()
+	err := w.store.db.Delete(options, key)
+	options.Close()
+	return err
 }
 
 // StartBatch start a new batch write processing
@@ -154,7 +96,9 @@ func (w *LVDBWriter) FlushBatch() error {
 	defer w.mutex.Unlock()
 
 	if w.store.writeBatch != nil {
-		err = w.store.db.Write(w.store.writeOptions, w.store.writeBatch)
+		options := defaultWriteOptions()
+		err = w.store.db.Write(options, w.store.writeBatch)
+		options.Close()
 		// After flush, release the writeBatch for future uses
 		w.store.writeBatch.Clear()
 		w.isBatch = false
